Added multiple theme support
