+++ app/src/main/java/be/ppareit/swiftp/gui/MainActivity.java
@@ -21,64 +21,37 @@ package be.ppareit.swiftp.gui;
 
 import android.Manifest;
 import android.app.AlertDialog;
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
-import android.content.SharedPreferences;
-import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
 import android.content.pm.PackageManager;
-import android.content.res.Resources;
-import android.net.Uri;
-import android.net.wifi.WifiConfiguration;
-import android.net.wifi.WifiManager;
 import android.os.Build;
 import android.os.Build.VERSION;
 import android.os.Bundle;
-import android.os.Handler;
-import android.preference.CheckBoxPreference;
-import android.preference.EditTextPreference;
-import android.preference.Preference;
-import android.preference.PreferenceActivity;
-import android.preference.PreferenceManager;
-import android.preference.PreferenceScreen;
-import android.preference.TwoStatePreference;
 import android.support.annotation.NonNull;
-import android.text.util.Linkify;
+import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.widget.BaseAdapter;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import net.vrallev.android.cat.Cat;
 
-import java.io.File;
-import java.net.InetAddress;
-import java.util.List;
-
-import be.ppareit.android.DynamicMultiSelectListPreference;
 import be.ppareit.swiftp.App;
 import be.ppareit.swiftp.BuildConfig;
-import be.ppareit.swiftp.FsService;
-import be.ppareit.swiftp.FsSettings;
 import be.ppareit.swiftp.R;
+import be.ppareit.swiftp.Util;
 
 /**
  * This is the main activity for swiftp, it enables the user to start the server service
  * and allows the users to change the settings.
  */
-public class MainActivity extends PreferenceActivity implements OnSharedPreferenceChangeListener {
-
-    private EditTextPreference mPassWordPref;
-    private Handler mHandler = new Handler();
+public class MainActivity extends AppCompatActivity{
 
     final static int PERMISSIONS_REQUEST_CODE = 12;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         Cat.d("created");
+        Util.setTheme(this);
         super.onCreate(savedInstanceState);
 
         if (VERSION.SDK_INT >= Build.VERSION_CODES.M) {
@@ -101,172 +74,9 @@ public class MainActivity extends PreferenceActivity implements OnSharedPreferen
             ad.show();
         }
 
-        addPreferencesFromResource(R.xml.preferences);
-
-        final SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(this);
-        Resources resources = getResources();
-
-        TwoStatePreference runningPref = findPref("running_switch");
-        updateRunningState();
-        runningPref.setOnPreferenceChangeListener((preference, newValue) -> {
-            if ((Boolean) newValue) {
-                startServer();
-            } else {
-                stopServer();
-            }
-            return true;
-        });
-
-        PreferenceScreen prefScreen = findPref("preference_screen");
-        Preference marketVersionPref = findPref("market_version");
-        marketVersionPref.setOnPreferenceClickListener(preference -> {
-            // start the market at our application
-            Intent intent = new Intent(Intent.ACTION_VIEW);
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            intent.setData(Uri.parse("market://details?id=be.ppareit.swiftp"));
-            try {
-                // this can fail if there is no market installed
-                startActivity(intent);
-            } catch (Exception e) {
-                Cat.e("Failed to launch the market.");
-                e.printStackTrace();
-            }
-            return false;
-        });
-        if (App.isFreeVersion() == false) {
-            prefScreen.removePreference(marketVersionPref);
-        }
-
-        updateLoginInfo();
-
-        EditTextPreference usernamePref = findPref("username");
-        usernamePref.setOnPreferenceChangeListener((preference, newValue) -> {
-            String newUsername = (String) newValue;
-            if (preference.getSummary().equals(newUsername))
-                return false;
-            if (!newUsername.matches("[a-zA-Z0-9]+")) {
-                Toast.makeText(MainActivity.this,
-                        R.string.username_validation_error, Toast.LENGTH_LONG).show();
-                return false;
-            }
-            stopServer();
-            return true;
-        });
-
-        mPassWordPref = findPref("password");
-        mPassWordPref.setOnPreferenceChangeListener((preference, newValue) -> {
-            stopServer();
-            return true;
-        });
-
-        DynamicMultiSelectListPreference autoconnectListPref = findPref("autoconnect_preference");
-        autoconnectListPref.setOnPopulateListener(
-                pref -> {
-                    Cat.d("autoconnect populate listener");
-
-                    WifiManager wifiManager = (WifiManager) getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-                    List<WifiConfiguration> configs = wifiManager.getConfiguredNetworks();
-                    if (configs == null) {
-                        Cat.e("Unable to receive wifi configurations, bark at user and bail");
-                        Toast.makeText(this,
-                                R.string.autoconnect_error_enable_wifi_for_access_points,
-                                Toast.LENGTH_LONG)
-                                .show();
-                        return;
-                    }
-                    CharSequence[] ssids = new CharSequence[configs.size()];
-                    CharSequence[] niceSsids = new CharSequence[configs.size()];
-                    for (int i = 0; i < configs.size(); ++i) {
-                        ssids[i] = configs.get(i).SSID;
-                        String ssid = configs.get(i).SSID;
-                        if (ssid.length() > 2 && ssid.startsWith("\"") && ssid.endsWith("\"")) {
-                            ssid = ssid.substring(1, ssid.length() - 1);
-                        }
-                        niceSsids[i] = ssid;
-                    }
-                    pref.setEntries(niceSsids);
-                    pref.setEntryValues(ssids);
-                });
-
-        EditTextPreference portnum_pref = findPref("portNum");
-        portnum_pref.setSummary(sp.getString("portNum",
-                resources.getString(R.string.portnumber_default)));
-        portnum_pref.setOnPreferenceChangeListener((preference, newValue) -> {
-            String newPortnumString = (String) newValue;
-            if (preference.getSummary().equals(newPortnumString))
-                return false;
-            int portnum = 0;
-            try {
-                portnum = Integer.parseInt(newPortnumString);
-            } catch (Exception e) {
-            }
-            if (portnum <= 0 || 65535 < portnum) {
-                Toast.makeText(MainActivity.this,
-                        R.string.port_validation_error, Toast.LENGTH_LONG).show();
-                return false;
-            }
-            preference.setSummary(newPortnumString);
-            stopServer();
-            return true;
-        });
-
-        Preference chroot_pref = findPref("chrootDir");
-        chroot_pref.setSummary(FsSettings.getChrootDirAsString());
-        chroot_pref.setOnPreferenceClickListener(preference -> {
-            AlertDialog folderPicker = new FolderPickerDialogBuilder(this, FsSettings.getChrootDir())
-                    .setSelectedButton(R.string.select, path -> {
-                        if (preference.getSummary().equals(path))
-                            return;
-                        if (!FsSettings.setChrootDir(path))
-                            return;
-                        // TODO: this is a hotfix, create correct resources, improve UI/UX
-                        final File root = new File(path);
-                        if (!root.canRead()) {
-                            Toast.makeText(this,
-                                    "Notice that we can't read/write in this folder.",
-                                    Toast.LENGTH_LONG).show();
-                        } else if (!root.canWrite()) {
-                            Toast.makeText(this,
-                                    "Notice that we can't write in this folder, reading will work. Writing in subfolders might work.",
-                                    Toast.LENGTH_LONG).show();
-                        }
-
-                        preference.setSummary(path);
-                        stopServer();
-                    })
-                    .setNegativeButton(R.string.cancel, null)
-                    .create();
-            folderPicker.show();
-            return true;
-        });
-
-        final CheckBoxPreference wakelock_pref = findPref("stayAwake");
-        wakelock_pref.setOnPreferenceChangeListener((preference, newValue) -> {
-            stopServer();
-            return true;
-        });
-
-        Preference help = findPref("help");
-        help.setOnPreferenceClickListener(preference -> {
-            Cat.v("On preference help clicked");
-            Context context = MainActivity.this;
-            AlertDialog ad = new AlertDialog.Builder(context)
-                    .setTitle(R.string.help_dlg_title)
-                    .setMessage(R.string.help_dlg_message)
-                    .setPositiveButton(android.R.string.ok, null)
-                    .create();
-            ad.show();
-            Linkify.addLinks((TextView) ad.findViewById(android.R.id.message),
-                    Linkify.ALL);
-            return true;
-        });
-
-        Preference about = findPref("about");
-        about.setOnPreferenceClickListener(preference -> {
-            startActivity(new Intent(this, AboutActivity.class));
-            return true;
-        });
-
+        getFragmentManager().beginTransaction()
+                .replace(android.R.id.content, new PreferenceFragment())
+                .commit();
     }
 
     @Override
@@ -322,136 +132,4 @@ public class MainActivity extends PreferenceActivity implements OnSharedPreferen
 
         return true;
     }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-
-        updateRunningState();
-
-        Cat.d("onResume: Register the preference change listner");
-        SharedPreferences sp = getPreferenceScreen().getSharedPreferences();
-        sp.registerOnSharedPreferenceChangeListener(this);
-
-        Cat.d("onResume: Registering the FTP server actions");
-        IntentFilter filter = new IntentFilter();
-        filter.addAction(FsService.ACTION_STARTED);
-        filter.addAction(FsService.ACTION_STOPPED);
-        filter.addAction(FsService.ACTION_FAILEDTOSTART);
-        registerReceiver(mFsActionsReceiver, filter);
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-
-        Cat.v("onPause: Unregistering the FTPServer actions");
-        unregisterReceiver(mFsActionsReceiver);
-
-        Cat.d("onPause: Unregistering the preference change listner");
-        SharedPreferences sp = getPreferenceScreen().getSharedPreferences();
-        sp.unregisterOnSharedPreferenceChangeListener(this);
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sp, String key) {
-        updateLoginInfo();
-    }
-
-    private void startServer() {
-        sendBroadcast(new Intent(FsService.ACTION_START_FTPSERVER));
-    }
-
-    private void stopServer() {
-        sendBroadcast(new Intent(FsService.ACTION_STOP_FTPSERVER));
-    }
-
-    private void updateLoginInfo() {
-
-        String username = FsSettings.getUserName();
-        String password = FsSettings.getPassWord();
-
-        Cat.v("Updating login summary");
-        PreferenceScreen loginPreference = findPref("login");
-        loginPreference.setSummary(username + " : " + transformPassword(password));
-        ((BaseAdapter) loginPreference.getRootAdapter()).notifyDataSetChanged();
-
-        EditTextPreference usernamePref = findPref("username");
-        usernamePref.setSummary(username);
-
-        EditTextPreference passWordPref = findPref("password");
-        passWordPref.setSummary(transformPassword(password));
-    }
-
-    private void updateRunningState() {
-        Resources res = getResources();
-        TwoStatePreference runningPref = findPref("running_switch");
-        if (FsService.isRunning()) {
-            runningPref.setChecked(true);
-            // Fill in the FTP server address
-            InetAddress address = FsService.getLocalInetAddress();
-            if (address == null) {
-                Cat.v("Unable to retrieve wifi ip address");
-                runningPref.setSummary(R.string.running_summary_failed_to_get_ip_address);
-                return;
-            }
-            String iptext = "ftp://" + address.getHostAddress() + ":"
-                    + FsSettings.getPortNumber() + "/";
-            String summary = res.getString(R.string.running_summary_started, iptext);
-            runningPref.setSummary(summary);
-        } else {
-            runningPref.setChecked(false);
-            runningPref.setSummary(R.string.running_summary_stopped);
-        }
-    }
-
-    /**
-     * This receiver will check FTPServer.ACTION* messages and will update the button,
-     * running_state, if the server is running and will also display at what url the
-     * server is running.
-     */
-    BroadcastReceiver mFsActionsReceiver = new BroadcastReceiver() {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            Cat.v("action received: " + intent.getAction());
-            // remove all pending callbacks
-            mHandler.removeCallbacksAndMessages(null);
-            // action will be ACTION_STARTED or ACTION_STOPPED
-            updateRunningState();
-            // or it might be ACTION_FAILEDTOSTART
-            final TwoStatePreference runningPref = findPref("running_switch");
-            if (intent.getAction().equals(FsService.ACTION_FAILEDTOSTART)) {
-                runningPref.setChecked(false);
-                mHandler.postDelayed(
-                        () -> runningPref.setSummary(R.string.running_summary_failed),
-                        100);
-                mHandler.postDelayed(
-                        () -> runningPref.setSummary(R.string.running_summary_stopped),
-                        5000);
-            }
-        }
-    };
-
-    static private String transformPassword(String password) {
-        Context context = App.getAppContext();
-        SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(context);
-        Resources res = context.getResources();
-        String showPasswordString = res.getString(R.string.show_password_default);
-        boolean showPassword = showPasswordString.equals("true");
-        showPassword = sp.getBoolean("show_password", showPassword);
-        if (showPassword)
-            return password;
-        else {
-            StringBuilder sb = new StringBuilder(password.length());
-            for (int i = 0; i < password.length(); ++i)
-                sb.append('*');
-            return sb.toString();
-        }
-    }
-
-    @SuppressWarnings({"unchecked", "deprecation"})
-    protected <T extends Preference> T findPref(CharSequence key) {
-        return (T) this.findPreference(key);
-    }
-
 }
