+++ app/src/main/res/values/styles.xml
@@ -1,4 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="AppTheme" />
+    <!-- Stub themes for future modifications! -->
+    <style name="AppThemeDark" parent="Theme.AppCompat" />
+    <style name="AppThemeLight" parent="Theme.AppCompat.Light" />
+    <style name="AppThemeLight.DarkActionBar" parent="Theme.AppCompat.Light.DarkActionBar" />
 </resources>
